refactor: согласно ТЗ сделаны 3 Store & все данные для TrackersVC берутся из dataProvider
