+++ Tracker/CoreData/DataProvider.swift
@@ -1,5 +1,5 @@
 //
-//  NotepadStoreUpdate.swift
+//  DataProvider.swift
 //  Tracker
 //
 //  Created by Дмитрий Чалов on 12.12.2025.
@@ -43,82 +43,47 @@ final class DataProvider: NSObject {
     weak var delegate: DataProviderDelegate?
     
     private let context: NSManagedObjectContext
-    private let dataStore: TrackerDataStore
-    private lazy var fetchedResultsController: NSFetchedResultsController<TrackerCoreData> = {
-        let fetchRequest = NSFetchRequest<TrackerCoreData>(entityName: "TrackerCoreData")
-        
-        let predicate = getPredicateForCurrentDate()
-        fetchRequest.predicate = predicate
-        
-        fetchRequest.sortDescriptors = [
-            NSSortDescriptor(key: "category.title", ascending: true),
-            NSSortDescriptor(key: "name", ascending: true)
-        ]
-        
-        let fetchedResultsController = NSFetchedResultsController(
-            fetchRequest: fetchRequest,
-            managedObjectContext: context,
-            sectionNameKeyPath: "category.title",
-            cacheName: nil
-        )
-        
-        fetchedResultsController.delegate = self
-        try? fetchedResultsController.performFetch()
-        
-        return fetchedResultsController
-    }()
-
+    private let trackerStore: TrackerStore
+    private let trackerCategoryStore: TrackerCategoryStore
+    private let trackerRecordStore: TrackerRecordStore
+    
     private var insertedIndexes: IndexSet?
     private var deletedIndexes: IndexSet?
     
     private var currentDate: Date = Date()
-    private let uiColorMarshalling = UIColorMarshalling.shared
     
-    init(_ dataStore: TrackerDataStore, delegate: DataProviderDelegate) throws {
+    init(_ dataStore: DataStore, delegate: DataProviderDelegate) throws {
         guard let context = dataStore.managedObjectContext else {
             throw DataProviderError.failedToInitializeContext
         }
         self.delegate = delegate
         self.context = context
-        self.dataStore = dataStore
+        self.trackerStore = TrackerStore(context: context)
+        self.trackerCategoryStore = TrackerCategoryStore(context: context)
+        self.trackerRecordStore = TrackerRecordStore(context: context)
     }
     
     // Метод для получения предиката фильтрации по текущей дате
     private func getPredicateForCurrentDate() -> NSPredicate? {
-        logger.info("called: \(#function) \(#line)")
+        logger.info("called: \(#function)")
         guard let currentWeekDay = WeekDay.fromDate(currentDate) else {
             logger.error("❌ Не удалось определить день недели для даты: \(currentDate)")
             return NSPredicate(value: false)
         }
-        // Проблема: schedule хранится как Data, нельзя фильтровать через contains
-        // Поэтому фильтруем вручную в shouldDisplayTracke
-         return createComplexPredicate(for: currentWeekDay)
+        return createComplexPredicate(for: currentWeekDay)
     }
     
     private func createComplexPredicate(for weekDay: WeekDay) -> NSPredicate? {
-        logger.info("called: \(#function) \(#line)")
-        // Сложный предикат для фильтрации в CoreData
-        // Работает только если schedule хранится как String, а не Data
-        
-        // Конвертируем WeekDay в строку для поиска
+        logger.info("called: \(#function)")
         let dayString = "\(weekDay.rawValue)"
-        
-        // Ищем JSON строку, содержащую этот номер дня
-        // Пример: schedule содержит "[1,3,5]" - ищем "1" или ",1," или "[1" или "1]"
         return NSPredicate(format: "schedule CONTAINS %@", dayString)
     }
     
     // Обновить предикат при смене даты
     private func updatePredicate() {
-        logger.info("called: \(#function) \(#line)")
+        logger.info("called: \(#function)")
         let predicate = getPredicateForCurrentDate()
-        fetchedResultsController.fetchRequest.predicate = predicate
-        
-        do {
-            try fetchedResultsController.performFetch()
-        } catch {
-            print("❌ Ошибка обновления предиката: \(error)")
-        }
+        trackerStore.updateFetchedResultsControllerPredicate(predicate)
     }
 }
 
@@ -126,7 +91,7 @@ final class DataProvider: NSObject {
 extension DataProvider: DataProviderProtocol {
     func fetchCompletedRecords() -> [TrackerRecord] {
         logger.info("called: \(#function)")
-        return (try? dataStore.fetchRecords()) ?? []
+        return (try? trackerRecordStore.fetchAllRecords()) ?? []
     }
 
     func toggleRecord(trackerId: UUID, date: Date) {
@@ -141,9 +106,9 @@ extension DataProvider: DataProviderProtocol {
 
         do {
             if exists {
-                try dataStore.deleteRecord(trackerId: trackerId, date: day)
+                try trackerRecordStore.deleteRecord(trackerId: trackerId, date: day)
             } else {
-                try dataStore.addRecord(trackerId: trackerId, date: day)
+                try trackerRecordStore.addRecord(trackerId: trackerId, date: day)
             }
         } catch {
             print("❌ Ошибка toggleRecord: \(error)")
@@ -152,13 +117,12 @@ extension DataProvider: DataProviderProtocol {
 
     var numberOfCategories: Int {
         logger.info("called: \(#function)")
-        return fetchedResultsController.sections?.count ?? 0
+        return trackerStore.fetchedResultsController.sections?.count ?? 0
     }
     
     func numberOfTrackersInCategory(_ section: Int) -> Int {
         logger.info("called: \(#function)")
-        // ВАЖНО: Проверяем существование секции
-        guard let sections = fetchedResultsController.sections,
+        guard let sections = trackerStore.fetchedResultsController.sections,
               section < sections.count else {
             print("⚠️ Ошибка: запрошенной секции \(section) не существует")
             return 0
@@ -170,31 +134,28 @@ extension DataProvider: DataProviderProtocol {
     
     func tracker(at indexPath: IndexPath) -> TrackerCoreData? {
         logger.info("called: \(#function)")
-        // ВАЖНО: Проверяем валидность indexPath
-        guard let sections = fetchedResultsController.sections,
+        guard let sections = trackerStore.fetchedResultsController.sections,
               indexPath.section < sections.count,
               indexPath.row < sections[indexPath.section].numberOfObjects else {
             print("❌ Ошибка: indexPath \(indexPath) вне границ")
             return nil
         }
-        return fetchedResultsController.object(at: indexPath)
+        return trackerStore.fetchedResultsController.object(at: indexPath)
     }
     
     func categoryTitle(at index: Int) -> String {
         logger.info("called: \(#function)")
-        guard let sections = fetchedResultsController.sections,
+        guard let sections = trackerStore.fetchedResultsController.sections,
               index < sections.count else {
             print("⚠️ Секция \(index) не существует")
             return "Категория"
         }
         let sectionInfo = sections[index]
-        // 2. Получаем первый трекер в секции
         guard let objects = sectionInfo.objects as? [TrackerCoreData],
               let firstObject = objects.first else {
             print("⚠️ Секция \(index) пустая")
             return "Категория \(index + 1)"
         }
-        // 3. Получаем связанную категорию
         guard let categoryEntity = firstObject.category,
               let title = categoryEntity.title, !title.isEmpty
         else {
@@ -204,59 +165,22 @@ extension DataProvider: DataProviderProtocol {
         return title
     }
     
-    func addTracker(_ tracker: Tracker, to: String) throws {
+    func addTracker(_ tracker: Tracker, to categoryTitle: String) throws {
         logger.info("called: \(#function)")
-        try? dataStore.addTracker(tracker, to: "Важное")
+        let category = try trackerCategoryStore.findOrCreateCategory(withTitle: categoryTitle)
+        try trackerStore.addTracker(tracker, to: category)
     }
     
     func deleteRecord(at indexPath: IndexPath) throws {
         logger.info("called: \(#function)")
-        let record = fetchedResultsController.object(at: indexPath)
-        try? dataStore.delete(record)
+        let trackerToDelete = trackerStore.fetchedResultsController.object(at: indexPath)
+        try trackerStore.deleteTracker(trackerToDelete)
     }
     
     // Установить текущую дату и обновить фильтрацию
     func setCurrentDate(_ date: Date) {
         logger.info("called: \(#function)")
         self.currentDate = date
-        
         updatePredicate()
-    }
-}
-
-// MARK: - NSFetchedResultsControllerDelegate
-extension DataProvider: NSFetchedResultsControllerDelegate {
-    func controllerWillChangeContent(_ controller: NSFetchedResultsController<NSFetchRequestResult>) {
-        logger.info("called: \(#function) \(#line)")
-        insertedIndexes = IndexSet()
-        deletedIndexes = IndexSet()
-    }
-
-    func controllerDidChangeContent(_ controller: NSFetchedResultsController<NSFetchRequestResult>) {
-        logger.info("called: \(#function) \(#line)")
-        delegate?.didUpdate(NotepadStoreUpdate(
-                insertedIndexes: insertedIndexes!,
-                deletedIndexes: deletedIndexes!
-            )
-        )
-        insertedIndexes = nil
-        deletedIndexes = nil
-    }
-    
-    func controller(_ controller: NSFetchedResultsController<NSFetchRequestResult>, didChange anObject: Any, at indexPath: IndexPath?, for type: NSFetchedResultsChangeType, newIndexPath: IndexPath?) {
-        logger.info("called: \(#function) \(#line)")
-
-        switch type {
-        case .delete:
-            if let indexPath = indexPath {
-                deletedIndexes?.insert(indexPath.item)
-            }
-        case .insert:
-            if let indexPath = newIndexPath {
-                insertedIndexes?.insert(indexPath.item)
-            }
-        default:
-            break
-        }
     }
 }
